pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aimovie-cloud"
version = "1.0.0"
description = "AI-powered video narration generator using cloud services"
authors = [{name = "AIMovie Team", email = "team@aimovie.com"}]
license = {text = "MIT"}
readme = "README_CLOUD.md"
requires-python = ">=3.8"
keywords = ["ai", "video", "narration", "cloud", "tts", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "streamlit>=1.28.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "moviepy>=1.0.3",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "edge-tts>=6.1.0",
    "dashscope>=1.14.0",
    "openai>=1.3.0",
    "loguru>=0.7.0",
    "pydantic>=2.4.0",
    "tqdm>=4.66.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
cloud = [
    "aliyun-python-sdk-core>=2.13.0",
    "oss2>=2.18.0",
    "tencentcloud-sdk-python>=3.0.0",
    "baidu-aip>=4.16.0",
]

dev = [
    "pytest>=7.4.0",
    "black>=23.9.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/aimovie/aimovie-cloud"
Documentation = "https://aimovie.readthedocs.io/"
Repository = "https://github.com/aimovie/aimovie-cloud.git"
Issues = "https://github.com/aimovie/aimovie-cloud/issues"

[project.scripts]
aimovie-cloud = "src.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
